+++ README.md
@@ -5,8 +5,8 @@
 ### 1단계 - 자동 DI
 #### 생성자 주입 - 수동
 다음 문제를 해결한다.
-- [ ] 테스트하기 어렵다.
-- [ ] Repository 객체를 교체하기 위해 또다른 객체를 만들어 바꿔줘야 한다. 즉, ViewModel에 직접적인 변경사항이 발생한다.
+- [x] 테스트하기 어렵다.
+- [x] Repository 객체를 교체하기 위해 또다른 객체를 만들어 바꿔줘야 한다. 즉, ViewModel에 직접적인 변경사항이 발생한다.
 
 #### 생성자 주입 - 자동
 다음 문제를 해결한다.
