+++ app/src/main/java/woowacourse/shopping/domain/CartRepository.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.domain
+
+import woowacourse.shopping.model.Product
+
+interface CartRepository {
+    fun addCartProduct(product: Product)
+
+    fun getAllCartProducts(): List<Product>
+
+    fun deleteCartProduct(id: Int)
+}
